% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.fasta.R
\name{fasta.read}
\alias{fasta.read}
\title{Read fasta formate and convert it to data frame}
\usage{
fasta.read(virus.fasta,host.fasta)
}
\arguments{
\item{virus.fasta}{directory path to the virus fasta file.}

\item{host.fasta}{directory path to the host fasta file.}
}
\value{
A list with two data frames.
}
\description{
Read fasta formate and convert it to data frame
}
\note{
The list with two data.frames; the first one for virus DNA sequences and the second one for the host.
}
\examples{
\dontrun{
fasta <- fasta.read("virus.fasta", "host.fasta")
fasta.v <- fasta[[1]]
fasta.h <- fasta[[2]]
}
}
\author{
Ali Mostafa Anwar \email{ali.mo.anwar@std.agr.cu.edu.eg} and Mohmed Soudy \email{MohmedSoudy2009@gmail.com}
}
